requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
kaleido==0.2.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
playwright==1.45.0
plotly==5.22.0
psycopg2-binary==2.9.9
pyee==11.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
tenacity==8.0.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.8
mailru-im-bot
